refactor: 에러 메시지 Enum으로 관리
